video(tracking): cleanup unused code
